pow: Add pow/pow_with_context function
Add a function to raise a BigDecimal to an integer power.

The number of digits would blow up if we kept the exact value, so
we use the precision provided in Context parameter (or the default
value).

We need to keep additional precision during intermediate
computations though, as each multiplication may lose up to 2 digits.

For tests, we compare against values that we can obtain from
Wolfram Alpha, and test a few cases with 50 or 500 precision context,
to check that the precision loss compensation work we do works
correctly.

For now, only support positive exponent values.

Fixes #147.